pyproject.toml
[tool.poetry]
name = "test-task-forager-ai"
version = "0.1.0"
description = "This is test task for Junior Python Developer position"
authors = ["hifichevymane <blekbjek@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
wemake-python-styleguide = "^0.18.0"
black = "^23.12.1"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
